feat: update presentation/representers/location & mixin form add location Harvey
