+++ app/application/services/add_location.rb
@@ -5,8 +5,6 @@ require 'dry/transaction'
 module Leaf
   module Service
     # Service to handle adding a new location
-    # :reek:FeatureEnvy
-    # :reek:UncommunicativeVariableName
     class AddLocation
       include Dry::Transaction
 
@@ -17,18 +15,13 @@ module Leaf
 
       # Step 1: Validate input
       def validate_input(input)
-        if input[:location_query].to_s.strip.empty?
-          Failure('Validation failed: location_query cannot be empty')
-        else
-          Success(input)
-        end
-      rescue StandardError => e
-        Failure("Validation error: #{e.message}")
+        result = Leaf::Request::NewLocationRequest.new(input).call
+        result.to_monad # 確保返回 Success 或 Failure
       end
 
+      # Step 2: Fetch location from external service
       def fetch_location(input)
         location_query = CGI.unescape(input[:location_query])
-        # puts("Decoded Location Query: #{location_query}")
 
         mapper = Leaf::GoogleMaps::LocationMapper.new(
           Leaf::GoogleMaps::API,
@@ -37,8 +30,8 @@ module Leaf
 
         location_entity = mapper.find(location_query)
         location_entity ? Success(location_entity) : Failure('Location not found')
-      rescue StandardError => e
-        Failure("Error fetching location: #{e}")
+      rescue StandardError => error # rubocop:disable Naming/RescuedExceptionsVariableName
+        Failure("Error fetching location: #{error}")
       end
     end
   end
