+++ app/presentation/representers/location.rb
@@ -1,26 +1,19 @@
 # frozen_string_literal: true
 
-module Views
-  # View for a single location entity
-  class Location
-    def initialize(location_entity)
-      @location = location_entity
-    end
+require 'roar/decorator'
+require 'roar/json'
 
-    def name
-      @location.name # || 'N/A'
-    end
+module Leaf
+  module Representer
+    # Represents a Location entity
+    class Location < Roar::Decorator
+      include Roar::JSON
 
-    def longitude
-      @location.longitude # || 'N/A'
-    end
-
-    def latitude
-      @location.latitude # || 'N/A'
-    end
-
-    def plus_code
-      @location.plus_code # || 'N/A'
+      property :id
+      property :name
+      property :longitude
+      property :latitude
+      property :plus_code
     end
   end
 end
